+++ __tests__/unit/pseudo-element.spec.js
@@ -18,7 +18,7 @@ test("Pseudo element scss", t => {
   t.plan(3)
 
   postcss()
-    .use(stylelint({ code: invalidScss, config: config,}))
+    .use(stylelint({ code: invalidScss, config: config, quietDeprecationWarnings: true, }))
     .process(invalidScss, { syntax: scssSyntax })
     .then(checkResult)
     .catch(logError)