+++ __tests__/unit/qualifying-element.spec.js
@@ -26,7 +26,7 @@ test("Qualifying element scss", t => {
   t.plan(6)
 
   postcss()
-    .use(stylelint({ code: invalidScss, config: config,}))
+    .use(stylelint({ code: invalidScss, config: config, quietDeprecationWarnings: true, }))
     .process(invalidScss, { syntax: scssSyntax })
     .then(checkResult)
     .catch(logError)