+++ __tests__/unit/zero-unit.spec.js
@@ -14,7 +14,7 @@ test("Zero unit scss", t => {
   t.plan(2)
 
   postcss()
-    .use(stylelint({ code: invalidScss, config: config,}))
+    .use(stylelint({ code: invalidScss, config: config, quietDeprecationWarnings: true, }))
     .process(invalidScss, { syntax: scssSyntax })
     .then(checkResult)
     .catch(logError)